backend/pyproject.toml
[project]
name = "interfaces-backend"
version = "0.1.0"
description = "Backend API for DPEX"
requires-python = ">=3.10,<3.13"
license = { text = "Proprietary" }

dependencies = [
    "fastapi>=0.109.0",
    "python-multipart>=0.0.9",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.0.0",
    "docker>=7.1.0",
    "httpx>=0.27.0",
    "pyzmq>=26.2.1",
    "PyYAML>=6.0",
    "python-dotenv>=1.0.0",
    "pyarrow>=16.0.0",
    "rich>=13.7.0",
    "huggingface_hub>=0.23.0",
    "supabase>=2.0.0",
    "percus-ai[storage,training]",
    "lerobot[feetech]",
]

[project.scripts]
phi-server = "interfaces_backend.main:main"

[project.optional-dependencies]
dev = [
    "ruff",
    "pytest",
    "httpx",  # for testing
]
build = [
    "pyinstaller>=6.0.0",  # for Tauri sidecar binary
]
ssh = [
    "paramiko>=3.0.0",  # for SSH connections to training instances
]
cloud = [
    "paramiko>=3.0.0",
    "verda>=1.21.0",  # for Verda cloud management
    "boto3>=1.34.0",  # for S3/R2 storage
    "vastai-sdk>=0.5.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/interfaces_backend"]

[tool.uv.sources]
percus-ai = { path = "../../features", editable = true }
lerobot = { path = "../../features/lerobot", editable = true }
